Move DesktopIcon.ts to Electron.
Created a bunch of problems with transpiling now transpiling into dist-electron/electron and dist-electron/ui INSTEAD of transpiling directly into dist-electron.

npm run dev expects the transpiled code to just be in dist-electron. Thus it would transpile and not actually update the running code but the subfolders code, which was not what npm run dev is using therefore it would re-run old code and show no updates. There is probably some way to change this to look for the dist-electron/electron or /ui for the new transpiled code. But for simplicity and because I'm too much of a noob at electron it's much easier to just keep all the .ts files together, so when you transpile it keeps them in the root folder of dist-electron.