appdata-file:// Protocol and displaying icons from appdata
Added appdata-file:// protocol for accessing files in AppData/Roaming/AltDesktop directory (mostly just icon images).

Now all images must be loaded from via the appdata-file:// protocol. With either a full directory or a relative sub-directory to AppData/Roaming/AltDesktop.

See the readme for getting the icons/images to work. Eventually the icons will be editable but for now editing the .json manually and placing the images manually is the only way to test.