Can now use relative path for desktopIcons.json Images
Instead of saving the entire file path, we can just set it up to use the relative AppData/Roaming/AltDesktop paths. This would simplify saving the file path, and essentially make it "portable" as in you could copy over the AltDesktop file to another pc and it would not break the path due to a different username, as all paths are relative to the AltDesktop folder instead of the entire directory path.

Format for working path is "/folder_name/image.png" or "/image.png" if file is directly in "AppData/Roaming/AltDesktop".