Directly throw error for non-relative paths in appdata-file protocol
Since we are using relative paths and appdata-file:// protocol does not use external hosts. Any time a host is returned for a request it means that it is trying to load an external Non-AppData directory file. Thus we can immediately respond with 403 and log an error.

Previously absolutely nothing would trigger the security check, because we were appending appDataBasePath to the requested url and then just seeing if the full path started with appDataBasePath, which of course it does since we just appended it.....