Cleanup (remove ElectronResources charts + functions).
Anything similar to this will be created as separate widgets which would be placed inside of the DesktopGrid. So for now am removing all their functions and will rewrite them or repurpose the functions later on.

Removing to simplify everything and focus on the Desktop. Had already removed rendering of the Charts in the last commit as it was interfering with the placement of the contextMenus.