Converted to only using iconsMap.
Removed the need to use icons/setIcons, and thus the need to keep both up to date.

Icons are now only stored in a Map which means all set, get etc. should be O(1) time.

Also added a updateIconField([row,col], field, value) which easily updates fields of an icon.

Last commit where I converted it to a Map was slightly better but I had left setIcons usable and was updating that, which in turn was O(n) thus invalidating most of the progress from moving to a Map. This is now fixed.

Still ironing out some kinks to make it more efficient/understandable