Create 'data' folder instead of icons.
Because icons will be stored in a folder with their [row, col] reference as a name, ensure that this folder is created on launch.

Basically extremely silly to put all icons in a single file (icons folder was just used for quick tests to ensure it would work to load from appData).