ensureDataFolder to create /data/ directories.
The file system will be /AppData/Roaming/AltDesktop/data/[row, col]/ (icons go here).

Added ensureDataFolder( row, col) to ensure that the file directory exists and will be called when adding icons to the local directory.

For now it is only called in ensureAppDataFiles, which sets up all the base appDataFiles, once directly with (0, 0) just to ensure it works as expected.