Icons load from /data/[row,col] folder.
Important: Folders are now [row,col] (no space) because formatting can get messy with paths with spaces.

Now loads directly from the /data/[row,col]/ folder. I.E. no more saving /icons/image.png, just place the image in [row,col] folder and save the .json with /image.png for imagepath.

Known issue: bad file paths that do not end in .png or signify that he file is an image return nothing thus do not render an icon. This either needs to be special cased to backup load unknown.png or it needs a special case around the call to get the safeFileProtocol where if it returns nothing replace it with unknown.png