Test: Adjusting Opacity of VideoBackground.
Added a parameter "opacity" for VideoBackground component which now sets the opacity of the video.

All of the background stuff is still in development, I am moreso pushing it to the limits now when I don't have to worry about breaking a ton of stuff so that when it gets fully implemented later with settings and everything it is a lot smoother.

Note: No background will show if background.mp4 does not exist in root directory. The videobackground will be completely transparent.