Added a HoverOpacityItem component.
Basically the plan for this is a component which can be put on the desktop which when hovered would set the background/videobackground to be transparent. Allowing you to see through the window without alt+tabbing or minimizing etc.

Ideally the user would be able to place this item anywhere from the eventual settings menu.

For now just logs "hovered" to console. Will need to set up ipc communcations and a way to change the opacity of videoBackground dynamically.