HoverOpacityItem updates tranparency.
Updated HoverOpacityItem to when hovered, now changes the VideoBackground to become transparent, allowing you to see through the program. For now Icons and the header remains rendered and the only thing that becomes transparent is the VideoBackground.

Also realized the sendHoverAction might not have been needed when almost all uses will probably just be frontend implemented.... oh well it was a simple implementation anyways.