Export SafeImage and getImagePath to SafeImage.tsx
Exported the SafeImage react component to it's own file.

Also remove the JSDoc comment lines. I was under the impression that the top line of JSDoc comments were ignored from the actual peak of the definition, but either something changed in my environment or that was never the case, thus removed all JSDoc comment headers.