Moving mouse resets opacity timer.
Moving the mouse while within the HoverOpacityItem now resets the timer. Basically in order to trigger opacity now you must leave the mouse still within the HoverOpacityItem for 150ms. Thus it is much less likely to accidentally trigger the opacity when slowly dragging a mouse over the area (like when moving the mouse to the header actions/other screens etc.