Removed IPC for HoverItem
HoverItem does not currently use the ipc, it is all done frontend in UI. Looking forward, cannot think of specific reasons why this needs ipc communication so unless the needs arise for it, it will not be restored.

Also slightly increased the size of the HoverOpacityItem. (eventually this will be moved and size set via settings.)