setIconData ipcHandler (writing to .json)
Created a ipcMainHandle for "setIconData" this is sent with the row, col, and an Icon. This will place the icon into the .json either overwriting the existing icon at [row, col] or adding it if it does not exist.

Currently only called via console by the command:
await window.electron.setIconData(2, 1, {row: 2, col: 1, image: "icon.png", name: "New Icon",});