BrowserWindow test.
Added a quick BrowserWindow (opens on program launch). Which loads from the EditIcon.tsx file to create a very basic window. Still a bunch of configuring and work to do but a decent start.

Some known issues: window can be maximized and minimized. window can be resized (not necessarily an issue, but does not stretch to fill window yet, probably should move it to its own css file).
Creates a separate instance in taskbar, ideally it should appear always on top, and block interactions with main window till closed/saved.
Need to be able to pass it parameters like which DesktopIcon data to load etc.