ContextMenu "open ICON_NAME" now opens subWindow
Now when right click-> "open ICON_NAME" it sends the ipc communication to open the subWindow. It also sends the row and col to the EditIcon.

For now this row and col only appear in the "Editing icon at row {row}, column {col}" line.

Not 100% set on the way it will load the desktopIcon, Ideally I want it to load from the .json as the actual 100% stored data to avoid and possible mistakes. But I also don't want to add an ipc channel for just loading an icon if I can avoid it.