DesktopIcon offsetX, offsetY
Added a DesktopIcon key for offsetX and offsetY. Both store numbers and can be used to add relative padding to a DesktopIcon as it is being rendered.

For example if you have an icon at [1,1] then have an offsetX of 10 it places it at normal location of [1,1] + 10 pixels for X.

This technically can cause issues because you can offset icons into other Grid slots, but is very nice for customizing a desktop where you don't want a strict grid structure. Say for instance you want the icons to be in a more circular shape or curved. Adding offsets effectively allow you free placement in the desktop. While still maintaining an easily accessible backend storage based on row,col (icon root), and creating the default intuitive desktopGrid for low-customization users.

Basically the grid row,col will work as a root/home location. Each [row,col] is unique and can hold 1 icon. But that icon can be offset within its grid, or technically freely placed anywhere else.

The ability to change the offsetX, offsetY will probably come quite a bit later into development because it will be part of the whole suite of customization features and will probably be mouse drag and dropping the icon NOT simply inputting raw values. So development on that alone will take some time, and there is much more important things to implement immediately. The only way to set offsets now is through editing the .json file. Then in DevTools console call (await window.electron.reloadIcon(row,col)).