Reloadable icons.
User can now right click an icon to open context menu, then select "Reload Icon" to reload fresh from the .json file.

Also updated the function "reloadIcon" to instead be "handleIpcReloadIcon" to remind any dev that the reloadIcon is NOT meant to be called, it is instead called as result of a callback from an ipc call like "await window.electron.reloadIcon(row, col)".

You can easily test this via editing the .json and then reloading the icon.