Tracking down the failed Validation.
Traced the Error: Failed to fetch icon data error down to the util.ts validateEventFrame function. Basically the problem is that the subwindow itself is seen as a different even frame, therefore when it requests getDesktopIcon it is seen as malicious since it does not originate from the source (main window). So now that the error is identified, just needs to be rewritten to allow a subWindow to also call ipc functions.