validateEventFrame updated to support subWindows
subWindowManager now creates the subWindow itself instead of just managing it.

Upon creating a window it adds the window hash to an allowedUrls which tracks all allowed URLs for the program.

validateEventFrame now compares the senderFrame against this array of allowed URLs to validate the frame and allow the call.

known non-ideal things (not full issues):
Too much logging, will be reduced as all the logging for debugging is still there.
subWindows do not get removed from allowedUrls on closing the subWindow.