uiLogger wrapper + console logging.
Added /util/uiLogger.ts
This is now a wrapper which allows the use of logging just like the main program logging. Just import, create a logger and log to it like normal.

This uiLogger also passes calls console.log statements for the same message. So no need to do:
console.log(data);
logger.info(data);

Just do:
logger.info(data);

also supports the usual:
logger.info("This is a message", data);

One thing to note is that this will work with subWindows, but the subWindows console.logs go to the subWindows devTools not the mainWindows devTools.