Consolidated base grid invisible padding.
Instead of using 1 variable GRID_PADDING for the invisible padding, it now uses ICON_PAD_TOP and ICON_PAD_LEFT. Reminder in practice this only affects the root of all DesktopIcons locations, it does not affect the distance between icons.

Reason this is called invisible padding is because when adding actual padding to the DesktopGrid object, it blocks right click/contextMenu events. So instead of adding padding direct, it is added to the root of all DesktopIcons.

Also updated some documentation for what things are.