New Icon now opens editIconWindow
Right click -> New Icon now opens up an editIconWindow for the icon that would fit in that area, or the closest valid icon location. (For example a click in the padding area left/top of the DesktopIcons root, will get redirected to Minimum 0, or Maximum numRows/numCols.

known issue: Opening the editIconWindow for an icon that does not exist yet does not work as expected. It currently only works if the icon exists, otherwise will open the menu and give the error that it cannot find the icon at row, col. This will be adjusted.