ContextMenu snapping to nearest icon for editing.
Now when right-clicking to open a contextMenu on the desktop background, it will default to the nearest possible icon grid slot, and instead show the icon contextMenu with "Open Icon_name" instead of "New Icon". This is so that when you click inside of a grid slot with an existing icon, or outside of the grid area it shows you the icon which you would be editing instead of just showing you "New Icon" and then opening up the edit to an existing icon.