Fixed clickboxes for contextMenu grid alignment.
Fixed a problem where it was not adjusted by headerHeight, and therefore a click in the bottom right section of an icon grid, would instead show the icon below it. (Only happened in the bottom right because the clickbox of the icon.name would override the desktop click. And since the 64px icons do not fully fill the 100px width icon grid, the bottom section would have some area where this happened.