headerAdjustY(y) and gridAdjustY(y) functions.
Simplified two common coordinate conversions for height.
One function which adjusts for header height. (headerAdjustY(y)).
This is for when you convert a global position through an event into a position relative to only the normal program (like it has no header).

Second added one which adjust for header and Grid offsets (gridAdjustY(y)). This is used for when you need to base the coordinate system on the root of the grid object. Like when detecting what icon grid slot your event took place in.