Creating Data/[row,col] file on editIcon
Now calls to ensure that the data/[row,col]/    folder exists when selecting "Edit icon_name" or "New Icon" from the contextMenu.

This is a crucial step for saving icon shortcuts/copies to the data files when the EditIcon window eventually allows uploading images instead of just taking the local directory to the file.