subWindows stay on top, and minimize/restore.
subWindows now are set to always show on top. Ideally this would be for only showing on top of the AltDesktop window, but this seems like a pain to make work so for now subWindows show on top of all windows.

Also subWindow properly calls .hide() and .show() when mainWindow is minimized/restored.
.show()/.hide() is used because Electron adds some goofy animation when .minimize()/.restore() is called for smaller windows.