Selecting icon now sets path to local path.
When saving an icon to the [row,col] file it now returns the localized path to the file.

This makes it that when a user uses the select image button, it saves the icon to the [row,col] folder and loads just the filename to the image path. Thus when saved properly renders the icon image.