Logic rewrite for getRowColFromXY()
Simplified getRowColFromXY() to assume local coordinates are passed.

The contextMenu for how handleEditIcon() gets its rowCol is already using local so nothing changes there.
The handleDesktopRightClick() has been adjusted to get the adjustedY from headerAdjustY() so that it too can call with local coordinates.

Removed the gridAdjustY() as it is no longer needed as the ICON_ROOT_OFFSET_TOP is already adjusted for in getRowColFromXY.

And now when clicking on the lines:
Vertical lines: returns left icon
Horizontal lines: returns top icon.

When clicking on an icon/icon name allocated space it will still call that icon regardless of what icon grid location the click is in. So beware if you have oversized/offset icons which cross into other icon grid locations.