launchWebsite now launches website in default browser.
launchWebsite(row, col) now correctly opens the icon's websiteLink path in the default browser.

This also includes a check to make sure that the website matches the expected shell.openExternal format to open in a web browser. Basically shell expects https:// or http:// in order to launch it as a website. So if the user does not declare that directly append it to make sure that it launches as a website in the default browser. Not adding this check means that something like "github.com" will just open the windows explorer expecting that to be a file path instead of a website.