Added "Website Link" to Edit icons fields.
Added "Website Link" which corresponds to websiteLink in desktopIcons.json.

This should be fine enough, since it already has backend https:// appending if saved without the full website link. i.e. saving something like "github.com" will work as the launchWebsite auto appends if it lacks a https:// or http:// at the beginning. Which ensures that it launches as a web link in default browser.