getFilePath on dropping into EditIcon.tsx.
Took quite a while to find a way for this to work without nodeIntegration. Shoutout to Omar Mihilmy for his youtube video with 58 views which had a very simple way to get the file path from a dropped file without enabling nodeIntegration. Spent well over an hour trying different ways and the Files passed kept appearing as undefined and null.