getFileType ipcHandler.
Added a getFileType ipcHandler to get the filetype from filePath(string) using mime-types and return it.

We can use this to later compare to see if the filetype is an image. If the filetype dropped is an image it can then be saved and override the Image Path text field. Otherwise we will assume it is a program and just override the Program Path text field.