Drag and drop display in EditIcon.tsx.
Added a "Drop image or program here" css element to the EditIcon.tsx when dragging a file on top of the window.

Drag and drop was already implemented but this is just a visual cue to confirm for the user that they can drag and drop image or program files.