Icon highlight on contextMenu on icon home.
Now adds a green highlight box around the icon image if the icon's home grid is right clicked, or the icon is right clicked (as that still triggers).

Basically if a user ever has a high offset on an icon and they click the home box for that icon it should highlight both the home box and the existing icon so the user can easily see which icon cooresponds to the icon home.

Still need to do a bunch of work with maintaining/disabling the highlighted state from icons/icon homes based on what the user does with the contextMenu.