Add small delay to repeat keybind press (like holding buttons)
Holding the keybinds by default can quickly get out of hand and increase load/stress on the pc. It can get "behind" where it keeps minimizing/restoring for seconds after the key is released.

To prevent accidentally lagging out the computer as it struggles to keep up with the keys being held down, added a small debouce timer to only allow it to be called every 50ms. While this might still cause issues holding the keybinds it should be significantly less than when it is left unchecked. While still allowing for quickly double tapping it if say it is stuck behind another window.