ContextMenu on an icon -> Edit correctly opens icon.
Fixed a problem where right clicking on an icon would show "edit iconName" option, but upon clicking would use the default handleEditIcon which would then recalculate the icon to edit based on x/y of the contextMenu.

This now when clicking on an icon calls handleEditIcon with a row, col and avoids calculating which to edit as it already has a target.