Updated DesktopIcon interface and small cleanup
Made width, height, and both offsets optional. I imagine in 90% of  cases they will not be used, and can always add them later, no need to directly predefine them. If anything these will be used as overrides so it is directly inhibiting to predefine them as when implemented if it contains a width/height or offsets it will assume that it has been selected to override the default values.