Modified non-square aspect ratio Icon positioning.
Wanted to ensure that icons still maintained a basic layout where icons with distinctly non-square aspect ratios would not stand out that badly.

To accomplish this the icons always receive the same square box (unless height and width are directly specified). But their location in that box fits the general theme for their full icon counterparts.

For example:
Portrait icons: are center justified so that they do not appear entirely on the left side of their box.
Landscape icons: are positioned at the bottom of their box, such that the gap between the icons and their names are all the same.