WIP fix for unknown.png in Non-dev (pre-built mode).
Really not happy with this implementation for multiple reasons. It is way too convoluted, but as a fallback to a "technically" working version I am commiting this.

Basically this is a forced fallback, where it tries to get the file through safeFileProtocol but the file does not exist, thus it force returns unknown.png. I am going to spend some time simplifying this and fixing it.