" " and "none" options to not render icon. And slight improvements.
Adding an image path = " " or "none" Will now not render an icon's image, the icon clickbox will still exist and icon name will still exist, just will not have an image.

Secondly added some slight improvements to the fallback for non-image files to call with appdata-file:///unknown. This unknown is then selected for and safeFileProtocol knows to automatically return the unknown.png image from src/assets/.

This is also slightly better from the last implementation because it is a delineated special case and can't be overriden by creating a /assets/unknown.png file.

Still not super happy with it, considering just creating an assets folder, and copying unknown.png to it and using that as the home of the unknown.png file. Which would simplify some things but would still be slightly sloppy.