Protocol block traversal attempts.
I swear I implemented this block before, but now blocks traversal attempts when calling safeFileProtocol.
For example safe file protocol from an icon path defaults to AppData/Roaming/AltDesktop/data/[row,col]/ directory. If the user then set an icon like "../../../../icon.png" this could escape the base path allowed: (AppData/Roaming/AltDesktop).
So when it detects that the fullpath does not start with the appDataBasePath it returns as File Not Found and does not access the file which is outside the base path.

Returning as if file not found still allows us to send back a default unknown.png image if the file requested was an image.
And in the future if we use this for non-image types we can send back their defaults as well.