Reloading icons fully reloads the image.
Fixed an issue where if the file the icon originally received back was changed (i.e. deleted or replaced with another file with same name), reloading the icon would not re-request the image but would just use the already cached image. Thus would require a restart (or refresh desktop) to show the updated file.

Instead this now has a forceReload prop and can be force reloaded and returns an image with a timestamp to break the cached version.

Also to note, Refresh Desktop did not change as it already effectively reloaded the frontend of the program regardless.