Simplified handleReloadIcon
Since handleReloadIcon always ipc calls a function which will return a call to handleIpcReloadIcon, there is no need to deal with anything but doing the call.

Then in the handleIpcReloadIcon it can do force reload with a new timestamp, and update the iconsMap.

Reminder there are still distinct uses for both. For instance handleReloadIcon requests the back end to send data for the frontend to update (updated icon name, filepaths etc.).
Then the handleIpcReloadIcon updates the icon to have the up to date information and force reloads the SafeIcon to make sure that the image is also up to date.