Fix <div> used within <p> error.
Accidentally was using <div> tags within a <p> and was generating errors upon right clicking a desktopIcon. For the most part didn't actually break anything so pushed off fixing it for a bit while I was working on the reloads and protocol stuff, but now that that is done it has been fixed.