CSS/HeroIcons update
Added HeroIcons library for custom icons like a folderIcon for opening a FileDialog. Will use this more in the future as it is a nice  set of MIT licensed icons.

Adjusted css for EditIcon header to look better.

Swapped the "select icon" text for a FolderIcon from heroIcons and added css to customize it.

Slightly increased the default width of the EditIcon window as the folder/select Icon was getting slightly cut off.