Adding image ensures unique name, + "( )" support in icon paths
Now when saveIconImage is called it will not replace the file in the directory with the same name with the new file but instead increment until it finds a unique name.

Therefore if the user has an icon.png and uses a different icon.png it will become icon(1).png so as to not delete the existing icon.png.

Also fixed an encoding error which would not display an icon image if it had the characters "(" or ")".