SubWindow no longer alwaysOnTop.
Rewrite of subWindowManager such that the subwindow does not always appear on top of all programs/windows. While maintaining the functionality of always appearing on top of the mainWindow, even without blocking the main process.

Also updated the fileDialog for image paths to support more image types: "bmp", "svg", "webp".
Previously would still work for drag and drop, but would not show up in FileDialog.

Still some things to work out:
Calling EditIcon on another desktopIcon with a subWindow open closes the subWindow, but does not open the EditIcon window.

Closing (to tray) the main program does not hide the activeSubWindow.