openFileDialog IPC comand now supports non-image.
Added a type parameter we can send to the ipc command. This will allow us to set what files we are looking for, for example passing it "image" returns with the image file extension filter.
Passing it something else just changes the name (bottom right of File Dialog filter) and lets all files show up.