OpenInExplorer ipc call added.
Still basic and requires some special casing.
For instance it expects to be called with the full relative path i.e. "data/[0,0]/icon.png" for images.

Also need to hook it up to the UI and need to fill some special cases like when file does not exist, attempt parent folder etc.