Very basic smallMenu (for future errors/confirmations).
A very basic ipc command:
await window.electron.showTestSmallMenu("title", "message")

has been added which will create a smallMenu.tsx browserWindow with for now just a title and message, and a basic close button to close the menu.

Still super early in development, and needs a lot of changes both to allow for sending messages back, and for general control and even more for making it look like a proper window.