small typo fix. (VSCode cache capitalization after creation)
Fixed a problem that occurred due to VSCode/ and I believe nodejs cache not liking it when you create a file, then change ONLY the capitalization of the file. The cache still sees the filename as the old version (smallMenu.tsx) instead of the current file (SmallMenu.tsx). This occurs in VSCode where it only gives an error on importing the updated file name. And appears in electron build where it gives the same error that name already added and differs only in capitalization.

If you ever come across this error just restart VSCode and it fixes it because it re-builds the cache with the up to date filenames.