showSmallWindow ipc call now returns Promise<string> of button clicked.
Now returns the string of the button clicked when you call showSmallWindow like such:
await window.electron.showSmallWindow("title", "message", ["These", "Are", "Buttons"])

Which would create a smallWindow with the title, and a message, and 3 buttons.

Upon the user clicking any of the buttons it returns with a string of the button name i.e. 'These' or 'Are' or 'Buttons'.

Therefore you can now add a warning/error/confirmation button which receives it's response and can do conditional logic based on the response.

Also returns 'Close' if the header button is closed. (Might need to add a case where either the close button does not render or defaults to something...)