showSmallWindow function
Added uiUtil.ts with a simple function for doing the await window.electron.showSmallWindow() call.

Doing this in its own function so that it can catch errors and this allows us to create logging statments in general if we wanted say "calling showSmallWindow ipc call with: title, message, buttons" etc. which are then used for all showSmallWindow ipc calls.

Mostly so we can avoid using try and catch blocks in every use case.