Attempt to resolve imagePath for full paths on saving image.
If the user pastes, or types out the full file path for an image path, upon save it now attempts to saveIconImage and save the image and the image path like it does when dropping the file, or when selecting the file from the fileDialog.

This fixes a likely common error of a user just copy pasting a filepath and then being confused when it comes back as unknown.png.