Verify source file exists for saveIconImage and show error.
Now verifies that the source file exists before attempting to copy it to local directory.

Also added a showSmallWindow if the attempted save fails due to image path being a full path but not an image.

Also to note this only triggers when it seems like a full image path, i.e. C:/Users/User/Desktop/ will save, but always show unknown and not attempt this copy. But C:/Users/User/Desktop/image.png will attempt this then show correctly or show an error that file does not exist.