Unified EditIcon closing function.
By default the subWindowHeader was just sending the ipc call to close. Without the EditIcon.tsx ever being notified in case it wanted to do anything specific before closing.

Updated this to accept an onClose reference which then was created and unified to be called at the end of saving, or on clicking the subWindowHeader X button to close.