Closing subwindow now sends signal to hideHighlightBox().
One annoying thing was opening an editIcon window and finding a way to have the icon and iconBox highlighted, then upon closing the editIcon window, un highlight the box. So a simple solution to this is to just send a signal upon CLOSE_SUBWINDOW to hide the highlight. Which then gets received by the renderer which calls hideHighlightBox() to unhighlight the icon and iconBox.