If image already in Icon Image folder, skip copying.
Quick skip for when the user selects a file that is already in the data/[row,col]/  folder to not make another copy but just return the file name as it is already in the folder.

Odds are rare that this ever really gets used, but it feels weird to pass it the literal file already in the folder and have it create another copy.