Now live previews: "Image Path".
Added the live preview to Image Path.

Also fixed the smallWindow added to be shown on error from last commit to correctly only show when an actual error has occurred, not when it succeeds (Accidentally left it on true, instead of false, so I could see how it looked if it were to appear...).