Allow blank as a response to number input fields.
Now allows "" or blank to be considered default in the fields.

Basically how it works is if blank it sets it undefined, which then in DesktopGrid.tsx it uses defaults if the field is null or undefined. So blank fields use default.

Basically before if you changed the field at all, attempting to set it to blank would end up setting the field to 0. So it was actually impossible to go back to global default if ever changed.

Also to note, if left blank instead of just changing the desktopIcons.json to undefined or null etc., it just removes the line from the .json.