Fix closing without saving from focusing other programs.
Fixed a problem when closing both a SmallWindow and a SubWindow at the same time, causing focus to be shifted to the last non-AltDesktop program.

The reason this seems to arise is that the focus within the program gets confused when both windows are closed at the same time. For instance I believe that basically the SmallWindow closes and it redirects focus to the EditIcon (this is fine).
Then the EditIcon closes and it attempts to redirect focus to the last window, but the SmallWindow does not exist i.e. null. I think from there it just uses the default OS focus which will just bring the last focused non-same-program into focus.
Therefore it skips the mainWindow and brings focus to another program.
Basically the way it has been fixed is that we instead requestMainWindowFocus before closing the subWindow(EditIcon). Thus the last focus is not null and it can just go back to focusing the mainWindow.