Simplified last commit (window focus on closing multiple windows fix).
Realized the exact moment I pushed that, this was only going to apply to EditIcon and I would have to do this for Settings eventually. Quickly thought, I'll just override the CLOSE_SUBWINDOW command to call it....

OH.... wait all it does is focus the mainWindow, I can just add that line to the CLOSE_SUBWINDOW command....

So now maintains that the subWindow always has mainWindow as last focused when closing thus should avoid any case where it focuses another program on close. Avoids the need for an ipc command and now applies to closing all subWindows.