Prevent opening EditIcon while already editing an icon.
Spent a long time trying to create a system where clicking to open a new editIcon would just call the handleClose() for the already open editIconWindow. But this seemed extremely buggy.

At first it was seeing icon as null (due to only running the useEffect once at start), fixed that by setting a dependence to icon. Then ran into a bunch of problems with the smallWindow initializing but the window closing before it displayed or received a response...

Might look into doing this later, but for now just disabled opening a new Edit Icon if there is an activeSubWindow.