With open subWindow, right clicks now re-center subWindow.
Right clicking other icons and such can change the highlight off of the actual target for the edit icon menu. Thus right clicks on desktop or Icons are now ignored while the subWindow is active.

And since we are calling getSubWindowState() it re-centers them on call.

left clicks still work as defined sine they do not affect the highlight.