refactor getSubWindowState -> isSubWindowActive.
Originally made this isSubWindowActive ipc call a while ago, and was initially confused as to why I never had an ipc call that used it.

Found out it was being used, but was through a different call "isSubWindowActive".

Figured this better fit the usage so renamed the invoke to match. Also updated some logging/comments.