getSetting, Settings.tsx prep, and DesktopIcon refactor.
Added a getSetting which directly reads from the file and returns value for key, or defaultSettings value for key, or null.

Prepped Settings.tsx to maintain a settings useState of type SettingsData. Which will eventually be passed via ipc to be saved to file. For now it just logs the settings object when you click save.

Refactored DesktopIcon to be used in the types.d.ts, with as much as it gets used makes sense to just keep it as a direct type object which then does not have to be directly imported by every file that wants to use it.