Heavily WIP videoFileProtocol / non-local Background Path.
Early work in progress version of a video Background Path implementation.

Settings Background Path expects the full file path.

Known issues: Some videos do not work. This is due to a problem with how the file is being streamed.
It appears that ffmpeg movflags and faststart play a role in whether a video will play or not. (Not limited by size/bitrate/fps etc. I have gotten a 4k 60fps 20kbps bitrate video to work).

Partial manual solution: moving metadata to start via:
"ffmpeg -i background.mp4 -movflags +faststart -c copy background_fixed.mp4"

This also has some problems with reloading the window causing it to fail, or when it reaches the end and goes to reload it also might fail. So this is not a permanent solution.

I also tried a more native approach with a contextBridge in preload for loading the video files, it worked fine but it drastically increased the memory usage. Evidently it was not very efficient memorywise as the program which with this implementation only takes up 200-230mb of ram with a 4k video playing, while that was taking up close to 1gb. Kept the files just in case this has unsolvable issues, and perhaps it would be optimizable.