Massive videoBackground rework.
Fixed certain videos not loading correctly.
Fixed the same videos(modified) loading but failing on reloading window or upon looping.
Greatly optimized video playback performance using multiple APIs/buffer tweaks.
Retries on failing.
Added a lot of logging for tracking the video background(will later need simplification to avoid spamming).

Videos which in the last version would not load, or would have the above fixed problems technically still have problems with the video stream being suspended. This does not affect their playback or performance, but does spam the logs file if logging is set to verbose. Thinking that it is possible that this might cause a problem in a certain situation I have left a logger statement which will log every time a video is loaded and it reaches a suspendCount of 25 times. I.e. if the video suspends loading 25 times it logs a short debug statement.