Skip attempt to load video file if backgroundPath not video.
Having a backgroundpath that is not a video file will now skip attempting to load it as a video and will just set the video to null.

Also fixed the videoFileProtocol to only log performance metrics when video is actually requesting (has sent at least 1 request before). So for example a non-video background path would not post the metrics. (I still need to rework logging for video errors/information, might need to create a separate logger or have a toggle for more verbose logging).