Added Video loggerFile.
Added a video_ logger which creates another file inside AppData/AltDesktop/logs/  .

Added createVideoLoggerForFile(fileName), which now logs to that file on logger.info(""): etc.

Because videoBackground is constantly going and has situations where it can sometimes spam the logger, this has been added to separate out the video logs if there is a problem with it.

Need to implement a way to write to it from Ui still....